// src/app/services/iex/interfaces/stats.ts
export interface Stats {
  companyName: string;
  marketcap: Number;
  beta: Number;
  week52high: Number;
  week52low: Number;
  week52change: Number;
  shortInterest: Number;
  shortDate: string;
  dividendRate: Number;
  dividendYield: Number;
  exDividendDate: string;
  latestEPS: Number;
  latestEPSDate: string;
  sharesOutstanding: Number;
  float: Number;
  returnOnEquity: Number;
  consensusEPS: Number;
  numberOfEstimates: Number;
  symbol: string;
  EBITDA: Number;
  revenue: Number;
  grossProfit: Number;
  cash: Number;
  debt: Number;
  ttmEPS: Number;
  revenuePerShare: Number;
  revenuePerEmployee: Number;
  peRatioHigh: Number;
  peRatioLow: Number;
  EPSSurpriseDollar: Number;
  EPSSurprisePercent: Number;
  returnOnAssets: Number;
  returnOnCapital: Number;
  profitMargin: Number;
  priceToSales: Number;
  priceToBook: Number;
  day200MovingAvg: Number;
  day50MovingAvg: Number;
  institutionPercent: Number;
  insiderPercent: Number;
  shortRatio: Number;
  year5ChangePercent: Number;
  year2ChangePercent: Number;
  year1ChangePercent: Number;
  ytdChangePercent: Number;
  month6ChangePercent: Number;
  month3ChangePercent: Number;
  month1ChangePercent: Number;
  day5ChangePercent: Number;
}